web/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
asyncpg==0.30.0
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
pydantic[email]==2.10.3
passlib==1.7.4
python-multipart==0.0.20
python-docx==1.1.2
